refactor(web): extract Skills API + PrivacySection to satisfy file-size gate
Analytics instrumentation pushed api/client.ts (1588) and SettingsApp.tsx (1531)
over the 1500-LOC file-size limit. Split the cohesive Skills API into
api/skills.ts (re-exported from client.ts via `export * from` — circular-safe
since post/get/put are hoisted function declarations) and the Privacy &
Analytics settings panel into settings/PrivacySection.tsx. client.ts -> 1275,
SettingsApp.tsx -> 1428. No behavior change; importers unchanged.
